Added "Recently Used Avatars" list.

+++ ReModCE/UI/ReAvatarList.cs
@@ -38,6 +38,8 @@ namespace ReModCE.UI
         }
 
         private readonly UiAvatarList _avatarList;
+
+        private readonly bool _hasPagination;
         private readonly ReUiButton _refreshButton;
         private readonly ReUiButton _nextPageButton;
         private readonly ReUiButton _prevPageButton;
@@ -60,13 +62,15 @@ namespace ReModCE.UI
         private readonly string _title;
 
         private readonly IAvatarListOwner _owner;
-        public ReAvatarList(string title, IAvatarListOwner owner) : base(
+        public ReAvatarList(string title, IAvatarListOwner owner, bool addPagination = true) : base(
             LegacyAvatarList,
             LegacyAvatarList.transform.parent,
             $"{title}AvatarList")
         {
+            _hasPagination = addPagination;
             _owner = owner;
             _title = title;
+
             _avatarList = GameObject.GetComponent<UiAvatarList>();
             _avatarList.clearUnseenListOnCollapse = false;
             _avatarList.field_Public_EnumNPublicSealedvaInPuMiFaSpClPuLiCrUnique_0 =
@@ -74,48 +78,63 @@ namespace ReModCE.UI
             GameObject.transform.SetSiblingIndex(0);
 
             var enableDisableListener = GameObject.AddComponent<EnableDisableListener>();
-            enableDisableListener.OnEnableEvent += () => Refresh(_owner.GetAvatars());
+            enableDisableListener.OnEnableEvent += RefreshAvatars;
 
             var expandButton = GameObject.GetComponentInChildren<Button>(true);
             _textComponent = expandButton.GetComponentInChildren<Text>();
             Title = title;
 
-            _refreshButton = new ReUiButton("↻", new Vector3(980f, 0f),
-                new Vector2(0.25f, 1), () => Refresh(_owner.GetAvatars()), expandButton.transform);
-
-            _nextPageButton = new ReUiButton("→", new Vector2(900f, 0f), new Vector2(0.25f, 1f), () =>
+            if (_hasPagination)
             {
-                _currentPage += 1;
-                Refresh(_owner.GetAvatars());
-            }, expandButton.transform);
+                _refreshButton = new ReUiButton("↻", new Vector3(980f, 0f), new Vector2(0.25f, 1), RefreshAvatars, expandButton.transform);
 
-            _prevPageButton = new ReUiButton("←", new Vector2(750f, 0f), new Vector2(0.25f, 1f), () =>
-            {
-                _currentPage -= 1;
-                Refresh(_owner.GetAvatars());
-            }, expandButton.transform);
+                _nextPageButton = new ReUiButton("→", new Vector2(900f, 0f), new Vector2(0.25f, 1f), () =>
+                {
+                    _currentPage += 1;
+                    Refresh(_owner.GetAvatars());
+                }, expandButton.transform);
 
-            _pageCount = new ReUiText("0 / 0", new Vector2(825f, 0f), new Vector2(0.25f, 1f), expandButton.transform);
+                _prevPageButton = new ReUiButton("←", new Vector2(750f, 0f), new Vector2(0.25f, 1f), () =>
+                {
+                    _currentPage -= 1;
+                    Refresh(_owner.GetAvatars());
+                }, expandButton.transform);
+
+                _pageCount = new ReUiText("0 / 0", new Vector2(825f, 0f), new Vector2(0.25f, 1f), expandButton.transform);
+            }
+        }
+
+        private void RefreshAvatars()
+        {
+            Refresh(_owner.GetAvatars());
         }
 
         public void Refresh(AvatarList avatars)
         {
-            var pagesCount = avatars.Count / MaxAvatarsPerPage;
-            _currentPage = Mathf.Clamp(_currentPage, 0, pagesCount);
-
-            _pageCount.Text = $"{_currentPage + 1} / {pagesCount + 1}";
-            var cutDown = avatars.GetRange(_currentPage * MaxAvatarsPerPage, Math.Abs(_currentPage * MaxAvatarsPerPage - avatars.Count));
-            if (cutDown.Count > MaxAvatarsPerPage)
+            if (_hasPagination)
             {
-                cutDown.RemoveRange(MaxAvatarsPerPage, cutDown.Count - MaxAvatarsPerPage);
+                var pagesCount = avatars.Count / MaxAvatarsPerPage;
+                _currentPage = Mathf.Clamp(_currentPage, 0, pagesCount);
+
+                _pageCount.Text = $"{_currentPage + 1} / {pagesCount + 1}";
+                var cutDown = avatars.GetRange(_currentPage * MaxAvatarsPerPage,
+                    Math.Abs(_currentPage * MaxAvatarsPerPage - avatars.Count));
+                if (cutDown.Count > MaxAvatarsPerPage)
+                {
+                    cutDown.RemoveRange(MaxAvatarsPerPage, cutDown.Count - MaxAvatarsPerPage);
+                }
+
+                _prevPageButton.Interactable = _currentPage > 0;
+                _nextPageButton.Interactable = _currentPage < avatars.Count / MaxAvatarsPerPage;
+
+                Title = $"{_title} ({cutDown.Count}/{avatars.Count})";
+
+                _avatarList.StartRenderElementsCoroutine(cutDown);
             }
-
-            _prevPageButton.Interactable = _currentPage > 0;
-            _nextPageButton.Interactable = _currentPage < avatars.Count / MaxAvatarsPerPage;
-
-            Title = $"{_title} ({cutDown.Count}/{avatars.Count})";
-
-            _avatarList.StartRenderElementsCoroutine(cutDown);
+            else
+            {
+                _avatarList.StartRenderElementsCoroutine(avatars);
+            }
         }
     }
 }
